Couple of renames for clarity
