+++ optd-dsl/src/analyzer/mod.rs
@@ -1,5 +1,5 @@
 pub mod context;
-pub mod error;
+pub mod errors;
 pub mod from_ast;
 pub mod hir;
 pub mod map;
