+++ optd-dsl/src/cli/main.rs
@@ -28,7 +28,7 @@
 //! ```
 use clap::{Parser, Subcommand};
 use optd_dsl::compile::{CompileOptions, adt_check, ast_to_hir, parse, scope_check};
-use optd_dsl::utils::error::{CompileError, Diagnose};
+use optd_dsl::utils::errors::{CompileError, Diagnose};
 use std::error::Error;
 use std::fs;
 use std::path::PathBuf;
