+++ optd-dsl/src/utils/mod.rs
@@ -1,4 +1,4 @@
-pub mod error;
+pub mod errors;
 pub mod span;
 #[cfg(test)]
 pub mod tests;
